% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_district.R
\name{compute_district_114}
\alias{compute_district_114}
\title{Title}
\usage{
compute_district_114(data, scales, state = NULL)
}
\arguments{
\item{county}{}
}
\description{
Title
}
\examples{
library(dplyr)
districts114_flat |> rename(state = state) |> compute_district_114() |> head()
nc_flat |> rename(fips = FIPS) |> compute_county_nc(county = "Ashe")
}
