% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_district118.R
\name{geom_district_118}
\alias{geom_district_118}
\title{Title}
\usage{
geom_district_118(
  mapping = NULL,
  data = NULL,
  position = "identity",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  ...
)
}
\arguments{
\item{...}{}
}
\description{
Title
}
\examples{

library(ggplot2)
library(dplyr)
districts_118_flat |>
mutate(id = 1:n()) |>
ggplot() +
aes(state = STATE_NAME, district = DISTRICT) +
geom_district_118(keep_state = "New York") +
aes(fill = id)
}
