% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_district116.R
\name{compute_district_116}
\alias{compute_district_116}
\title{Title}
\usage{
compute_district_116(
  data,
  scales,
  keep_state = NULL,
  district = NULL,
  drop_state = NULL
)
}
\arguments{
\item{county}{}
}
\description{
Title
}
\examples{
library(dplyr)
#districts_116_flat |> rename(state = STATE_NAME, district = DISTRICT) |> compute_district_116() |> head()
}
